refactor: 입력값 파서 생성
