+++ src/main/kotlin/calculator/StringCalculator.kt
@@ -1,0 +1,8 @@
+package calculator
+
+object StringCalculator {
+    fun add(input: String?): Int {
+        val numbers = InputParser.parse(input)
+        return numbers.sum
+    }
+}
